+++ codegen/dict_validator.py
@@ -131,13 +131,15 @@ class MapValidator(Validator[Any, Dict[T1, T2], Serializable]):
             return Err({OBJECT_ERRORS_FIELD: [expected("a map")]})
 
 
-class IsDict(Validator[Any, Dict[Any, Any], Serializable]):
+class IsDictValidator(Validator[Any, Dict[Any, Any], Serializable]):
     def __call__(self, val: Any) -> Result[Dict[Any, Any], Serializable]:
         if isinstance(val, dict):
             return Ok(val)
         else:
             return Err({OBJECT_ERRORS_FIELD: [expected("a dictionary")]})
 
+
+is_dict_validator = IsDictValidator()
 
 def _has_no_extra_keys(
     keys: Set[str],
@@ -160,7 +162,7 @@ def _has_no_extra_keys(
 def _dict_without_extra_keys(
     keys: Set[str], data: Any
 ) -> Result[Dict[Any, Any], Serializable]:
-    return IsDict()(data).flat_map(_has_no_extra_keys(keys))
+    return is_dict_validator(data).flat_map(_has_no_extra_keys(keys))
